RavenDB-3808 InitializeTopology and InitializeEmptyTopologyWithId don't have to be static. Fixing the condition when we force engine to be a candidate after we initialize new cluster.
